feat: AddDetailTripPlanDto 클래스 PK 재설정

+++ src/main/java/com/ssafy/tourmates/client/tripPlan/service/dto/AddDetailTripPlanDto.java
@@ -8,10 +8,10 @@ import java.util.List;
 @Data
 public class AddDetailTripPlanDto {
 
-    private List<Long> contendIds;
+    private List<Integer> contendIds;
 
     @Builder
-    public AddDetailTripPlanDto(List<Long> contendIds) {
+    public AddDetailTripPlanDto(List<Integer> contendIds) {
         this.contendIds = contendIds;
     }
 }
